grid sweep cannot be average grid color, but rather last grid color collectec from the two for loop method
i need to only compare the last one, since the first ones are of the ocean and not of a potenitally black screen

however, fishing time is not constant, unlike what i previously thought. need a way to see when the fishing stops, quickly do a grid sweep for like 1 or half a second, and then move on.

for if we get a boot or trash or things that will cause the fishing game to end but the program may still keep trying to find 100%